+++ game/cards/repository.go
@@ -395,6 +395,7 @@ var DM05 = map[string]match.CardConstructor{
 	"60b50334-f4c9-46dd-aa87-76f4e67cb5f8": dm05.LeQuistTheOracle,
 	"181126d5-0ad8-4033-a9d2-c8ca939f3fac": dm05.SteelTurretCluster,
 	"1064bbcc-0f15-4757-8d88-93084edaba3f": dm05.Gigakail,
+	"cfedb8c9-419a-4c0b-b05b-a3a30436f4b0": dm05.WispHowlerShadowOfTears,
 }
 
 var DM06 = map[string]match.CardConstructor{